requirements.txt
streamlit
cryptography
pillow
